requirements.txt
python-dotenv==1.2.2
requests==2.32.5
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
